feat: Support programatically wrapping lambda handlers
