+++ instrumentation/aws_lambda/lib/opentelemetry/instrumentation/aws_lambda/instrumentation.rb
@@ -21,6 +21,7 @@ module OpenTelemetry
         private
 
         def require_dependencies
+          require_relative 'wrap'
           require_relative 'handler'
         end
       end
